+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file. The project
 
 ## [Unreleased]
 
+### Breaking changes
+- The `StorageKey` trait is re-implemented for signed integer types (`i8`, `i16`, `i32` and `i64`)
+  to achieve the natural ordering of produced keys. This change will break indices using
+  signed integers as keys. To emulate the old implementation, you may create a wrapper
+  around a type (e.g., `struct QuirkyI32Key(i32)`) and implement `StorageKey` for it
+  using big endian encoding. Then, use the wrapper instead of the int type in indices.
+  See the unit tests for `StorageKey` for an example.
+
 ## 0.5 - 2018-01-30
 
 ### Breaking changes